Create 9 button 3*3 grid
Create an HTML page with a paragraph written on it and under 
which 9 buttons are placed in a 3X3 grid. The first row is for 
buttons labeled with colors names Red, Green, and Blue, the 
second row with numbers 10, 20, 30, and the third row with 
different font names. Click event of each of the buttons should 
make the appropriate change in the style of paragraph.