Create pet name,age,json
Create a form that takes data about a pet. The form must be well 
designed and should accept the pet’s name, age, weight, type, 
and what it likes most. At the submission of this form create a 
Pet object in JavaScript filled with these values and log that 
object and equivalent JSON on the console.